+++ src/utils/lf-localization.service.ts
@@ -1,4 +1,5 @@
 import { formatString } from './string-utils.js';
+import { LfLanguageCookie, getLfLanguageCookie } from './cookie-utils.js';
 
 export type resourceType = { language: string; resource: object };
 
@@ -67,7 +68,10 @@ export class LfLocalizationService implements ILocalizationService {
       console.log('No static resources provided. Call initResourcesFromUrlAsync to load resources dynamically.');
     }
     try {
-      this.setLanguage(window?.navigator?.language ?? this.DEFAULT_LANGUAGE);
+      const domainCookie: string = document?.cookie ?? '';
+      const cookieLanguage: LfLanguageCookie | undefined = getLfLanguageCookie(domainCookie);
+      const language = cookieLanguage?.uic ?? window?.navigator?.language ?? this.DEFAULT_LANGUAGE;
+      this.setLanguage(language);
     } catch {
       this.setLanguage(this.DEFAULT_LANGUAGE);
     }
@@ -102,31 +106,33 @@ export class LfLocalizationService implements ILocalizationService {
    * @param url
    */
   private async getSelectedLanguageResourceAsync(url: string) {
-      try {
-        await this.trySetLanguageResourceAsync(url, this._selectedLanguage);
-        return;
-      } catch (e) {
-        console.warn(`Selected language resource ${this._selectedLanguage} is not found at ${url}${this._selectedLanguage}.json.`);
-        if ((e as Error)?.name === ResourceNotFoundError_NAME) {
-          const closestLanguage: string = this.mapToClosestLanguage(this._selectedLanguage);
-            try {
-              await this.trySetLanguageResourceAsync(url, closestLanguage);
-              return;
-            } catch {
-              console.warn(`Language resource ${closestLanguage} is not found at ${url}${closestLanguage}.json.`);
-            }
-        } else {
-          console.error(e);
+    try {
+      await this.trySetLanguageResourceAsync(url, this._selectedLanguage);
+      return;
+    } catch (e) {
+      console.warn(
+        `Selected language resource ${this._selectedLanguage} is not found at ${url}${this._selectedLanguage}.json.`
+      );
+      if ((e as Error)?.name === ResourceNotFoundError_NAME) {
+        const closestLanguage: string = this.mapToClosestLanguage(this._selectedLanguage);
+        try {
+          await this.trySetLanguageResourceAsync(url, closestLanguage);
+          return;
+        } catch {
+          console.warn(`Language resource ${closestLanguage} is not found at ${url}${closestLanguage}.json.`);
         }
+      } else {
+        console.error(e);
+      }
     }
     this.setLanguageResource(this.DEFAULT_LANGUAGE);
   }
 
   private async trySetLanguageResourceAsync(url: string, language: string): Promise<void> {
     if (!this.setLanguageResource(language)) {
-        await this.addResourceFromUrlAsync(`${url}${language}.json`, language);
-        this.setLanguageResource(language);
-        console.log(`Loaded resource from ${url}${language}.json.`);
+      await this.addResourceFromUrlAsync(`${url}${language}.json`, language);
+      this.setLanguageResource(language);
+      console.log(`Loaded resource from ${url}${language}.json.`);
     }
   }
 
@@ -297,7 +303,7 @@ export class LfLocalizationService implements ILocalizationService {
     return false;
   }
 
-  private mapToClosestLanguage(originalLanguage: string) : string {
+  private mapToClosestLanguage(originalLanguage: string): string {
     const languageWithoutDash = this._selectedLanguage.split('-')[0];
     switch (languageWithoutDash) {
       case 'zh':